Remove duplicates from Alexa data

+++ tools/update-alexa-data.sh
@@ -15,17 +15,23 @@
 #
 #
 #
-# This script is provided to update the data/top-10000.txt file. 
+# This script is provided to update the data/top-10000.txt file.
 
-ALEXA_URL=http://s3.amazonaws.com/alexa-static/top-1m.csv.zip
+CSV='top-1m.csv'
+ALEXA_URL=http://s3.amazonaws.com/alexa-static/$CSV.zip
 TOP_COUNT=10000
-OUTPUT=top-${TOP_COUNT}.txt
+UNIQ_COUNT=12000
+OUTPUT=alexa-top-${TOP_COUNT}-global.txt
 
-echo "Fetching $ALEXA_URL"
-curl -O $ALEXA_URL
-unzip -o top-1m.csv.zip top-1m.csv
-head -$TOP_COUNT top-1m.csv | cut -d, -f2 | cut -d/ -f1 > $OUTPUT
-rm -f top-1m.csv*
+if [ ! -f "$CSV" ]; then
+  if [ ! -f "${CSV}.zip" ]; then
+    echo "${CSV}.zip not found - Fetching $ALEXA_URL"
+    curl -O $ALEXA_URL
+  fi
+  unzip -o $CSV.zip $CSV
+fi
+
+cut -d, -f2 $CSV | cut -d/ -f1 | head -$UNIQ_COUNT | ./ordered-uniq.py | head -$TOP_COUNT > $OUTPUT
 ls -la $OUTPUT
